requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
asgiref=3.8.1=pypi_0
attrs=24.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.0=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.18=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecated=1.2.14=pypi_0
durationpy=0.9=pypi_0
fastapi=0.115.4=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.22.0=pypi_0
google-api-python-client=2.151.0=pypi_0
google-auth=2.36.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.8.3=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.67.1=pypi_0
grpcio-status=1.67.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httplib2=0.22.0=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.26.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
kubernetes=31.0.0=pypi_0
langchain=0.3.7=pypi_0
langchain-community=0.3.5=pypi_0
langchain-core=0.3.15=pypi_0
langchain-google-genai=2.0.4=pypi_0
langchain-text-splitters=0.3.2=pypi_0
langsmith=0.1.142=pypi_0
libffi=3.4.4=hd77b12b_1
markdown-it-py=3.0.0=pypi_0
marshmallow=3.23.1=pypi_0
mdurl=0.1.2=pypi_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.20.0=pypi_0
openssl=3.0.15=h827c3e9_0
opentelemetry-api=1.28.1=pypi_0
opentelemetry-exporter-otlp-proto-common=1.28.1=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.28.1=pypi_0
opentelemetry-instrumentation=0.49b1=pypi_0
opentelemetry-instrumentation-asgi=0.49b1=pypi_0
opentelemetry-instrumentation-fastapi=0.49b1=pypi_0
opentelemetry-proto=1.28.1=pypi_0
opentelemetry-sdk=1.28.1=pypi_0
opentelemetry-semantic-conventions=0.49b1=pypi_0
opentelemetry-util-http=0.49b1=pypi_0
orjson=3.10.11=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pip=24.2=py311haa95532_0
posthog=3.7.0=pypi_0
propcache=0.2.0=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.28.3=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.6.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyreadline3=3.5.4=pypi_0
python=3.11.10=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py311haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.35=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.41.2=pypi_0
sympy=1.13.3=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.20.3=pypi_0
tqdm=4.67.0=pypi_0
typer=0.13.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.32.0=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
watchfiles=0.24.0=pypi_0
websocket-client=1.8.0=pypi_0
websockets=13.1=pypi_0
wheel=0.44.0=py311haa95532_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h8cc25b3_1
yarl=1.17.1=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h8cc25b3_1
